requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0